requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==24.3.0
audioread==3.0.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
decorator==5.1.1
distro==1.9.0
filelock==3.16.1
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.12.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
kiwisolver==1.4.8
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.44.0
MarkupSafe==3.0.2
matplotlib==3.10.0
more-itertools==10.6.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
networkx==3.4.2
nltk==3.9.1
numba==0.61.0
numpy==1.26.4
openai==0.28.0
openai-whisper==20240930
packaging==24.2
pillow==11.1.0
platformdirs==4.3.6
pooch==1.8.2
propcache==0.2.1
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
regex==2024.11.6
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.1
six==1.17.0
sniffio==1.3.1
soundfile==0.13.0
soxr==0.5.0.post1
sympy==1.13.3
textblob==0.19.0
threadpoolctl==3.5.0
tiktoken==0.8.0
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
yarl==1.18.3
